+++ app/assets/javascripts/angular/directives/work_packages/options-dropdown-directive.js
@@ -38,7 +38,8 @@ angular.module('openproject.workPackages.directives')
   'groupingModal',
   'QueryService',
   '$window',
-  function(I18n, columnsModal, exportModal, saveModal, settingsModal, shareModal, sortingModal, groupingModal, QueryService, $window){
+  '$state',
+  function(I18n, columnsModal, exportModal, saveModal, settingsModal, shareModal, sortingModal, groupingModal, QueryService, $window, $state){
 
   return {
     restrict: 'AE',
@@ -47,9 +48,6 @@ angular.module('openproject.workPackages.directives')
       angular.element($window).bind('click', function() {
         scope.$emit('hideAllDropdowns');
       });
-
-      // Modals
-      scope.showExportModal   = exportModal.activate;
 
       scope.saveQuery = function(){
         if(scope.query.isNew()){
@@ -69,11 +67,12 @@ angular.module('openproject.workPackages.directives')
             .then(function(data){
               settingsModal.deactivate();
               scope.$emit('flashMessage', data.status);
-              scope.$emit('queryResetRequired');
+              $state.go('work-packages.list', { query_id: null }, { reload: true });
             });
         }
       };
 
+      // Modals
       scope.showSaveAsModal = function(event){
         showExistingQueryModal.call(saveModal, event);
       };
@@ -86,6 +85,10 @@ angular.module('openproject.workPackages.directives')
         showExistingQueryModal.call(settingsModal, event);
       };
 
+      scope.showExportModal = function(){
+        showModal.call(exportModal);
+      };
+
       scope.showColumnsModal = function(){
         showModal.call(columnsModal);
       };
@@ -96,7 +99,7 @@ angular.module('openproject.workPackages.directives')
 
       scope.showSortingModal = function(){
         showModal.call(sortingModal);
-      }
+      };
 
       scope.toggleDisplaySums = function(){
         scope.$emit('hideAllDropdowns');