+++ lib/redmine.rb
@@ -108,6 +108,7 @@ Redmine::AccessControl.map do |map|
                                          :journals => [:index, :diff],
                                          :queries => [:index, :create, :update, :available_columns, :custom_field_filters, :grouped],
                                          :work_packages => [:show, :index],
+                                         :work_packages_api => [:get],
                                          :'work_packages/reports' => [:report, :report_details],
                                          :planning_elements => [:index, :all, :show, :recycle_bin],
                                          :planning_element_journals => [:index]}
@@ -294,10 +295,13 @@ Redmine::MenuManager.map :project_menu do |menu|
                       :if => Proc.new { |p| p.shared_versions.any? },
                       :html => {:class => "icon2 icon-process-arrow1"}
 
-  menu.push :work_packages, { controller: '/work_packages', action: 'index', set_filter: 1 },
+  menu.push :work_packages, { controller: '/work_packages', action: 'index' },
                             param: :project_id,
                             caption: :label_work_package_plural,
-                            html: {class: "icon2 icon-copy"}
+                            html: {
+                              class: "icon2 icon-copy",
+                              "data-ui-route" => ''
+                            }
 
   menu.push :new_work_package, { :controller => '/work_packages', :action => 'new'},
                                :param => :project_id,