+++ doc/CHANGELOG.md
@@ -76,6 +76,7 @@ See doc/COPYRIGHT.rdoc for more details.
 * `#7533` Fixes: Deleted block not addable
 * `#7562` Regression: creating ticket via API fails with HTTP 422
 * `#7608` Make highlight buttons styleable
+* `#7700` No flash message on status page
 * `#8228` Icons in select2 elements missing
 * `#8277` Fix arrow of column headers not restrained to one line
 * Allowed sending of mails with only cc: or bcc: fields